Add ruby command pattern
Use the command_demo_as_spec as the runner and integration test,
skip tests for the individual commands (thought I was going to
, hence the spec for command_shared_examples which is only
used by the macro_command_spec) but decided it was tedious.
Really, all of this was pretty tedious. But, got some good ruby mine
experience, and perhaps instilled a thing or two.