Add print_where to iterate through all items
Pass it a proc for the condition that must be passed. Keep it if it passed, leave nil
if it fails, calls it again if it's a menu.

Touched up the rest and tweaked it and the output to keep somewhat consistent with
what we did on the java side.